Created the 3 main files to be used
art_api.py: Sends requests to the Art Institute of Chicago api and processes the artwork retrieved. Will also create a color palette image based on the artwork.

twitter_api.py: Authentication for the twitter bot and tweets two images:
    1. The original artwork
    2. The color palette of the artwork
        obtained after processing
The caption will include details of the artwork and link to the artwork on the AIC's website.

main: General workflow of the bot which will include the other two files.