Caption and payload changes
Length of the title is now based on the 280 character limit of tweets and how many available characters are left after considering the other information shown in the tweet. The title will use the remaining characters and if it exceeds the limit, then it will be cut off.

About half of the query results are "asian art" so I changed the limit to 3 and added a random factor to decide which of the 3 artworks will be returned. This has shown a better chance of getting different types of artworks.